moving drain info and custodian under module as suggested from review
