+++ promotion/datastore.go
@@ -17,7 +17,6 @@ import (
 	errorutils "github.com/brave-intl/bat-go/utils/errors"
 	"github.com/brave-intl/bat-go/utils/jsonutils"
 	"github.com/brave-intl/bat-go/utils/logging"
-	"github.com/brave-intl/bat-go/utils/responses"
 	walletutils "github.com/brave-intl/bat-go/utils/wallet"
 	"github.com/getsentry/sentry-go"
 	"github.com/lib/pq"
@@ -121,7 +120,7 @@ type Datastore interface {
 	// GetDrainPoll gets the information about a drain poll job
 	GetDrainPoll(drainID *uuid.UUID) (*DrainPoll, error)
 	// GetCustodianDrainInfo gets the information about a drain poll job
-	GetCustodianDrainInfo(paymentID *uuid.UUID) ([]responses.CustodianDrain, error)
+	GetCustodianDrainInfo(paymentID *uuid.UUID) ([]CustodianDrain, error)
 }
 
 // ReadOnlyDatastore includes all database methods that can be made with a read only db connection
@@ -152,7 +151,7 @@ type ReadOnlyDatastore interface {
 	// GetDrainPoll gets the information about a drain poll job
 	GetDrainPoll(drainID *uuid.UUID) (*DrainPoll, error)
 	// GetCustodianDrainInfo gets the information about a drain poll job
-	GetCustodianDrainInfo(paymentID *uuid.UUID) ([]responses.CustodianDrain, error)
+	GetCustodianDrainInfo(paymentID *uuid.UUID) ([]CustodianDrain, error)
 }
 
 // Postgres is a Datastore wrapper around a postgres database
@@ -686,8 +685,8 @@ func (pg *Postgres) SaveClaimCreds(creds *ClaimCreds) error {
 }
 
 // GetCustodianDrainInfo Get the status of the custodian drain info
-func (pg *Postgres) GetCustodianDrainInfo(paymentID *uuid.UUID) ([]responses.CustodianDrain, error) {
-	resp := []responses.CustodianDrain{}
+func (pg *Postgres) GetCustodianDrainInfo(paymentID *uuid.UUID) ([]CustodianDrain, error) {
+	resp := []CustodianDrain{}
 	// get the linked wallet info
 	stmt := `
 select
@@ -697,7 +696,7 @@ from
 where
 	id = $1
 `
-	var custodian responses.Custodian
+	var custodian Custodian
 	if err := pg.RawDB().Get(&custodian, stmt, paymentID); err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, nil
@@ -721,7 +720,7 @@ where
 	wallet_id = $1
 `
 	type batchedPromotionsDrained struct {
-		responses.PromotionDrained
+		DrainInfo
 		BatchID uuid.UUID `db:"batch_id"`
 	}
 
@@ -733,20 +732,20 @@ where
 		return nil, err
 	}
 
-	batches := map[uuid.UUID][]responses.PromotionDrained{}
+	batches := map[uuid.UUID][]DrainInfo{}
 	batchValue := map[uuid.UUID]decimal.Decimal{}
 
 	// chunk all these into related batches
 	for i := 0; i < len(promosDrained); i++ {
 		if _, ok := batches[promosDrained[i].BatchID]; !ok {
-			batches[promosDrained[i].BatchID] = []responses.PromotionDrained{}
+			batches[promosDrained[i].BatchID] = []DrainInfo{}
 		}
 		if _, ok := batchValue[promosDrained[i].BatchID]; !ok {
 			batchValue[promosDrained[i].BatchID] = decimal.Zero
 		}
 		batches[promosDrained[i].BatchID] = append(
 			batches[promosDrained[i].BatchID],
-			responses.PromotionDrained{
+			DrainInfo{
 				PromotionID:   promosDrained[i].PromotionID,
 				TransactionID: promosDrained[i].TransactionID,
 				CompletedAt:   promosDrained[i].CompletedAt,
@@ -761,7 +760,7 @@ where
 	// for each batch go through and create a custodian drain and add to resp drain
 	// add values along the way
 	for k := range batches {
-		resp = append(resp, responses.CustodianDrain{
+		resp = append(resp, CustodianDrain{
 			BatchID:           k,
 			Custodian:         custodian,
 			PromotionsDrained: batches[k],
